Split NKModel.swift into designated source code files for every contained type.
