+++ Sources/NextcloudKit/Models/NKEditorDetailsCreators.swift
@@ -1,0 +1,15 @@
+// SPDX-FileCopyrightText: Nextcloud GmbH
+// SPDX-FileCopyrightText: 2019 Marino Faggiana
+// SPDX-FileCopyrightText: 2023 Claudio Cambra
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+import Foundation
+
+public class NKEditorDetailsCreators: NSObject {
+    public var editor = ""
+    public var ext = ""
+    public var identifier = ""
+    public var mimetype = ""
+    public var name = ""
+    public var templates: Int = 0
+}
