+++ Sources/NextcloudKit/Models/NKEditorTemplates.swift
@@ -1,0 +1,15 @@
+// SPDX-FileCopyrightText: Nextcloud GmbH
+// SPDX-FileCopyrightText: 2019 Marino Faggiana
+// SPDX-FileCopyrightText: 2023 Claudio Cambra
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+import Foundation
+
+public class NKEditorTemplates: NSObject {
+    public var delete = ""
+    public var ext = ""
+    public var identifier = ""
+    public var name = ""
+    public var preview = ""
+    public var type = ""
+}
