+++ Sources/NextcloudKit/Models/NKFileProperty.swift
@@ -1,0 +1,13 @@
+// SPDX-FileCopyrightText: Nextcloud GmbH
+// SPDX-FileCopyrightText: 2019 Marino Faggiana
+// SPDX-FileCopyrightText: 2023 Claudio Cambra
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+import Foundation
+
+public class NKFileProperty: NSObject {
+    public var classFile: String = ""
+    public var iconName: String = ""
+    public var name: String = ""
+    public var ext: String = ""
+}
