+++ Sources/NextcloudKit/Models/NKRichdocumentsTemplate.swift
@@ -1,0 +1,15 @@
+// SPDX-FileCopyrightText: Nextcloud GmbH
+// SPDX-FileCopyrightText: 2019 Marino Faggiana
+// SPDX-FileCopyrightText: 2023 Claudio Cambra
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+import Foundation
+
+public class NKRichdocumentsTemplate: NSObject {
+    public var delete = ""
+    public var ext = ""
+    public var name = ""
+    public var preview = ""
+    public var templateId: Int = 0
+    public var type = ""
+}
